refactor: moved transformations into new folder
